+++ test/demos/Demo3.java
@@ -5,10 +5,10 @@ import java.util.List;
 
 import org.eclipse.rdf4j.query.TupleQuery;
 import org.eclipse.rdf4j.query.TupleQueryResult;
+import org.eclipse.rdf4j.repository.Repository;
 
 import com.fluidops.fedx.Config;
 import com.fluidops.fedx.FedXFactory;
-import com.fluidops.fedx.FederationManager;
 import com.fluidops.fedx.QueryManager;
 import com.fluidops.fedx.structures.Endpoint;
 import com.fluidops.fedx.util.EndpointFactory;
@@ -23,7 +23,7 @@ public class Demo3 {
 		endpoints.add( EndpointFactory.loadSPARQLEndpoint("http://dbpedia", "http://dbpedia.org/sparql"));
 		endpoints.add( EndpointFactory.loadSPARQLEndpoint("http://swdf", "http://data.semanticweb.org/sparql"));
 		
-		FedXFactory.initializeFederation(endpoints);
+		Repository repo = FedXFactory.initializeFederation(endpoints);
 		
 		String q = "PREFIX rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>\n"
 			+ "PREFIX dbpedia-owl: <http://dbpedia.org/ontology/>\n"
@@ -32,13 +32,14 @@ public class Demo3 {
 			+ "?President dbpedia-owl:party ?Party . }";
 		
 		TupleQuery query = QueryManager.prepareTupleQuery(q);
-		TupleQueryResult res = query.evaluate();
+		try (TupleQueryResult res = query.evaluate()) {
 		
-		while (res.hasNext()) {
-			System.out.println(res.next());
+			while (res.hasNext()) {
+				System.out.println(res.next());
+			}
 		}
 		
-		FederationManager.getInstance().shutDown();
+		repo.shutDown();
 		System.out.println("Done.");
 		System.exit(0);