+++ test/demos/Demo6.java
@@ -4,9 +4,9 @@ import java.util.Collections;
 
 import org.eclipse.rdf4j.query.TupleQuery;
 import org.eclipse.rdf4j.query.TupleQueryResult;
+import org.eclipse.rdf4j.repository.Repository;
 
 import com.fluidops.fedx.FedXFactory;
-import com.fluidops.fedx.FederationManager;
 import com.fluidops.fedx.QueryManager;
 import com.fluidops.fedx.structures.Endpoint;
 
@@ -17,20 +17,21 @@ public class Demo6 {
 		
 		// the fedx config implicitly defines a dataConfig
 		String fedxConfig = "examples/fedxConfig-withPrefixDecl.prop";
-		FedXFactory.initializeFederation(fedxConfig, Collections.<Endpoint>emptyList());
+		Repository repo = FedXFactory.initializeFederation(fedxConfig, Collections.<Endpoint>emptyList());
 		
 		String q = "SELECT ?President ?Party WHERE {\n"
 			+ "?President rdf:type dbpedia:President .\n"
 			+ "?President dbpedia:party ?Party . }";
 		
 		TupleQuery query = QueryManager.prepareTupleQuery(q);
-		TupleQueryResult res = query.evaluate();
+		try (TupleQueryResult res = query.evaluate()) {
 		
-		while (res.hasNext()) {
-			System.out.println(res.next());
+			while (res.hasNext()) {
+				System.out.println(res.next());
+			}
 		}
 		
-		FederationManager.getInstance().shutDown();
+		repo.shutDown();
 		System.out.println("Done.");
 		System.exit(0);