+++ test/demos/QueryPlanLogDemo.java
@@ -7,7 +7,6 @@ import org.eclipse.rdf4j.repository.sail.SailRepository;
 
 import com.fluidops.fedx.Config;
 import com.fluidops.fedx.FedXFactory;
-import com.fluidops.fedx.FederationManager;
 import com.fluidops.fedx.monitoring.QueryPlanLog;
 
 public class QueryPlanLogDemo {
@@ -27,19 +26,21 @@ public class QueryPlanLogDemo {
 			+ "?President dbpedia-owl:party ?Party . }";
 		
 		TupleQuery query = repo.getConnection().prepareTupleQuery(QueryLanguage.SPARQL, q);
-		TupleQueryResult res = query.evaluate();
+		try (TupleQueryResult res = query.evaluate()) {
 		
-		int count = 0;
-		while (res.hasNext()) {
-			res.next();
-			count++;
+			int count = 0;
+			while (res.hasNext()) {
+				res.next();
+				count++;
+			}
+
+			System.out.println("# Done, " + count + " results");
 		}
-		System.out.println("# Done, " + count + " results");
 		
 		System.out.println("# Optimized Query Plan:");
 		System.out.println(QueryPlanLog.getQueryPlan());
 		
-		FederationManager.getInstance().shutDown();
+		repo.shutDown();
 		System.out.println("Done.");
 		System.exit(0);